Moved types to separate files
