+++ Sources/Endpoints/Endpoint.swift
@@ -43,11 +43,14 @@ public protocol DecoderType {
 
 extension JSONDecoder: DecoderType { }
 
-/// The `Response` is an associated type which defines the response from the server. Note that this is just type information which helpers, such as the built-in `URLSession` extensions, can use to know how to handle particular types. For instance, if this type conforms to `Decodable`, then a JSON decoder is used on the data coming from the server. If it's typealiased to `Void`, then the extension can know to ignore the response. If it's `Data`, then it can deliver the response data unmodified.
-/// An `ErrorResponse` type can be associated to define what value conforming to `Decodable` to use when parsing an error response from the server. This can be useful if your server returns a different JSON structure when there's an error versus a success. Often in a project, this can be defined globally and `typealias` can be used to associate this global type on all `Endpoint`s.
-/// 
+/// The `Response` is an associated type which defines the response from the server. Note that this is just type information which helpers, such as the built-in
+/// `URLSession` extensions, can use to know how to handle particular types. For instance, if this type conforms to `Decodable`, then a JSON decoder is used
+/// on the data coming from the server. If it's typealiased to `Void`, then the extension can know to ignore the response. If it's `Data`, then it can deliver the
+/// response data unmodified.
 ///
-///
+/// An `ErrorResponse` type can be associated to define what value conforming to `Decodable` to use when parsing an error response from the server.
+/// This can be useful if your server returns a different JSON structure when there's an error versus a success. Often in a project, this can be defined globally
+/// and `typealias` can be used to associate this global type on all `Endpoint`s.
 public protocol Endpoint {
     associatedtype Response
     associatedtype ErrorResponse: Decodable = EmptyResponse
@@ -60,6 +63,9 @@ public protocol Endpoint {
     associatedtype BodyEncoder: EncoderType = JSONEncoder
     associatedtype ErrorDecoder: DecoderType = JSONDecoder
     associatedtype ResponseDecoder: DecoderType = JSONDecoder
+
+    /// A `Definition` which pieces together all the components defined in the endpoint.
+    static var definition: Definition<Self> { get }
 
     /// The instance of the associated `Body` type. Must be `Encodable`.
     var body: Body { get }
@@ -82,128 +88,6 @@ public protocol Endpoint {
 
     /// The decoder instance to use when decoding the associated `Response` type
     static var responseDecoder: ResponseDecoder { get }
-
-    static var definition: Definition<Self> { get }
-}
-
-extension Endpoint {
-
-    /// Generates a `URLRequest` given the associated request value.
-    /// - Parameter environment: The environment in which to create the request
-    /// - Throws: An `EndpointError` which describes the error filling in data to the associated `Definition`.
-    /// - Returns: A `URLRequest` ready for requesting with all values from `self` filled in according to the associated `Endpoint`.
-    public func urlRequest(in environment: EnvironmentType) throws -> URLRequest {
-
-        var components = URLComponents()
-        components.path = Self.definition.path.path(with: pathComponents)
-
-        let urlQueryItems: [URLQueryItem] = try Self.definition.parameters.compactMap { item in
-
-            let value: Any
-            let name: String
-            switch item {
-            case .query(let queryName, let valuePath):
-                value = parameters[keyPath: valuePath]
-                name = queryName
-            case .queryValue(let queryName, let queryValue):
-                value = queryValue
-                name = queryName
-            default:
-                return nil
-            }
-
-            guard let queryValue = value as? ParameterRepresentable else {
-                throw EndpointError.invalidQuery(named: name, type: type(of: value))
-            }
-
-            if let encodedValue = queryValue.parameterValue {
-                return URLQueryItem(name: name, value: encodedValue)
-            }
-
-            return nil
-        }
-
-        let bodyFormItems: [URLQueryItem] = try Self.definition.parameters.compactMap { item in
-
-            let value: Any
-            let name: String
-            switch item {
-            case .form(let formName, let valuePath):
-                value = parameters[keyPath: valuePath]
-                name = formName
-            case .formValue(let formName, let formValue):
-                value = formValue
-                name = formName
-            default:
-                return nil
-            }
-
-            guard let formValue = value as? ParameterRepresentable else {
-                throw EndpointError.invalidForm(named: name, type: type(of: value))
-            }
-
-            if let encodedValue = formValue.parameterValue {
-                return URLQueryItem(name: name, value: encodedValue)
-            }
-
-            return nil
-        }
-
-        if !urlQueryItems.isEmpty {
-            components.queryItems = urlQueryItems
-        }
-
-        let baseUrl = environment.baseUrl
-        guard let url = components.url(relativeTo: baseUrl) else {
-            throw EndpointError.invalid(components: components, relativeTo: baseUrl)
-        }
-
-        var urlRequest = URLRequest(url: url)
-        urlRequest.httpMethod = Self.definition.method.methodString
-
-        let headerItems: [String: String] = try Self.definition.headers.reduce(into: [:]) { allHeaders, field in
-            let value: Any
-            let name = field.key.name
-
-            switch field.value {
-            case .field(let valuePath):
-                value = headerValues[keyPath: valuePath]
-            case .fieldValue(let fieldValue):
-                value = fieldValue
-            }
-
-            guard let headerValue = value as? CustomStringConvertible else {
-                throw EndpointError.invalidHeader(named: name, type: type(of: value))
-            }
-
-            allHeaders[name] = headerValue.description
-        }
-
-        for (name, value) in headerItems {
-            urlRequest.setValue(value, forHTTPHeaderField: name)
-        }
-
-        urlRequest.url = url
-
-        if !(body is EmptyResponse) {
-            do {
-                urlRequest.httpBody = try Self.bodyEncoder.encode(body)
-            } catch {
-                throw EndpointError.invalidBody(error)
-            }
-
-            if headerItems[Headers.contentType.name] == nil {
-                urlRequest.addValue("application/json", forHTTPHeaderField: Headers.contentType.name)
-            }
-        } else if !bodyFormItems.isEmpty {
-            urlRequest.httpBody = bodyFormItems.formString.data(using: .utf8)
-            urlRequest.setValue("application/x-www-form-urlencoded", forHTTPHeaderField: Headers.contentType.name)
-        }
-
-        urlRequest = environment.requestProcessor(urlRequest)
-
-        return urlRequest
-    }
 }
 
 public extension Endpoint where Body == EmptyResponse {
@@ -220,33 +104,6 @@ public extension Endpoint where Parameters == Void {
 
 public extension Endpoint where HeaderValues == Void {
     var headerValues: HeaderValues { return () }
-}
-
-/// The HTTP Method
-public enum Method {
-    case options
-    case get
-    case head
-    case post
-    case put
-    case patch
-    case delete
-    case trace
-    case connect
-
-    public var methodString: String {
-        switch self {
-        case .options: return "OPTIONS"
-        case .get: return "GET"
-        case .head: return "HEAD"
-        case .post: return "POST"
-        case .put: return "PUT"
-        case .patch: return "PATCH"
-        case .delete: return "DELETE"
-        case .trace: return "TRACE"
-        case .connect: return "CONNECT"
-        }
-    }
 }
 
 public extension Endpoint where ResponseDecoder == JSONDecoder {
@@ -266,17 +123,6 @@ public extension Endpoint where BodyEncoder == JSONEncoder {
     }
 }
 
-public protocol EnvironmentType {
-    /// The baseUrl of the Environment
-    var baseUrl: URL { get }
-    /// Processes the built URLRequest right before sending in order to attach any Environment related authentication or data to the outbound request
-    var requestProcessor: (URLRequest) -> URLRequest { get }
-}
-
-public extension EnvironmentType {
-    var requestProcessor: (URLRequest) -> URLRequest { return { $0 } }
-}
-
 public struct Definition<T: Endpoint> {
 
     /// The HTTP method of the Endpoint
@@ -286,7 +132,7 @@ public struct Definition<T: Endpoint> {
     /// The parameters (form and query) that are included in the Definition
     public let parameters: [Parameter<T.Parameters>]
     /// The headers that are included in the Definition
-    public let headers: [Headers: HeaderField<T.HeaderValues>]
+    public let headers: [Header: HeaderField<T.HeaderValues>]
 
     /// Initializes a Definition with the given properties, defining all dynamic pieces as type-safe parameters.
     /// - Parameters:
@@ -294,23 +140,13 @@ public struct Definition<T: Endpoint> {
     ///   - path: The path template representing the path and all path-related parameters
     ///   - parameters: The parameters passed to the endpoint. Either through query or form body.
     ///   - headerValues: The headers associated with this request
-    public init(method: Method, path: PathTemplate<T.PathComponents>, parameters: [Parameter<T.Parameters>] = [], headers: [Headers: HeaderField<T.HeaderValues>] = [:]) {
+    public init(method: Method,
+                path: PathTemplate<T.PathComponents>,
+                parameters: [Parameter<T.Parameters>] = [],
+                headers: [Header: HeaderField<T.HeaderValues>] = [:]) {
         self.method = method
         self.path = path
         self.parameters = parameters
         self.headers = headers
-    }
-}
-
-extension Array where Element == URLQueryItem {
-
-    /// Goes through each URLQueryItem element and joins them with a '&',
-    /// suitable for putting into the httpBody of a request
-    public var formString: String {
-        return map { item in
-            let name = item.name.pathSafe
-            let value = item.value?.pathSafe ?? ""
-            return "\(name)=\(value)"
-        }.joined(separator: "&")
     }
 }
